backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Contract Intelligence Parser Backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.5.3",
    "coverage>=7.10.5",
    "fastapi>=0.116.1",
    "pdfplumber>=0.11.7",
    "pymongo>=4.14.1",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "redis>=6.4.0",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ruff>=0.12.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
start-server = "uvicorn src.main:app --reload --host 0.0.0.0 --port 8000"
start-celery = "python -m celery -A src.tasks.celery worker --loglevel=info --pool=solo"
